+++ src/wlnFP/fingerprint.h
@@ -2,13 +2,14 @@
 #ifndef FINGERPRINT_H
 #define FINGERPRINT_H
 
-#include <inttypes.h>
+#include <stdlib.h>
+
 #define FPSIZE 32
 
-uint16_t* WLNFingerprint(const char* string);
+u_int8_t* WLNFingerprint(const char* string);
 bool WLNDescriptors(const char *string); 
 
-double WLNFPTanimoto(uint16_t *fp1, uint16_t *fp2); 
+double WLNFPTanimoto(u_int8_t *fp1, u_int8_t *fp2); 
 double OBabelTanimoto(const char *str1, const char *str2);
 
 #endif
